This is the singleton monitor
I have updated the monitor so that it is now functioning as a singleton, the bash script will need an update (i think) because now the monitor has to be called through a different method. I could be wrong tho. Client also has to be updated to fit singleton pattern.